Feature: WebSocketServer inside permanent scripts
