+++ server/providers/actions/scripts/apis/memebox.api.ts
@@ -1,117 +1,15 @@
 import {Service} from "@tsed/di";
 import {ActionTriggerEventBus} from "../../action-trigger-event.bus";
-import {
-  MediaType,
-  ScreenMediaOverridableProperties,
-  TriggerActionOverrides,
-  TriggerClipOrigin,
-  VisibilityEnum
-} from "@memebox/contracts";
+import {MediaType} from "@memebox/contracts";
 import {ActionActiveState} from "../../action-active-state";
-import {timeoutAsync} from "./sleep.api";
 import {Inject} from "@tsed/common";
 import {PERSISTENCE_DI} from "../../../contracts";
 import {Persistence} from "../../../../persistence";
-import merge from 'lodash/merge';
 import {DisposableBase} from "./disposableBase";
 import {takeUntil} from "rxjs/operators";
+import {ActionApi} from "./action.api";
+import {MediaApi} from "./media.api";
 
-
-export class ActionApi {
-  constructor(
-    protected memeboxApi: MemeboxApi,
-    protected actionId: string,
-    protected screenId?: string | undefined
-  ) {
-
-  }
-
-  async trigger(overrides?: TriggerActionOverrides | undefined): Promise<void> {
-    // if there is one already running
-
-    // todo add the current visibility it was trigggered with
-    // - to prevent waiting if its currently in toggle mode
-
-    // await this.memeboxApi.actionActiveState.waitUntilDoneAsync(this.actionId, this.screenId);
-
-    this.memeboxApi.actionTriggerEventBus.triggerMedia({
-      id: this.actionId,
-      origin: TriggerClipOrigin.Scripts,
-      originId: this.memeboxApi.scriptId,
-      targetScreen: this.screenId,
-      overrides
-    });
-
-    await this.memeboxApi.actionActiveState.waitUntilDoneAsync(this.actionId, this.screenId);
-  }
-}
-
-export class MediaApi extends ActionApi {
-  constructor(
-    memeboxApi: MemeboxApi,
-    actionId: string,
-    screenId?: string|undefined
-  ) {
-    super(memeboxApi, actionId, screenId);
-  }
-
-  updateScreenOptions(overrides: ScreenMediaOverridableProperties): Promise<void> {
-    this.memeboxApi.actionTriggerEventBus.updateMedia({
-      id: this.actionId,
-      origin: TriggerClipOrigin.Scripts,
-      originId: this.memeboxApi.scriptId,
-      targetScreen: this.screenId,
-      overrides: {
-        screenMedia: overrides
-      }
-    })
-
-    return timeoutAsync(50);
-  }
-
-  async triggerWhile(executionFunction: (
-    helpers:
-    {
-      reset: () => void,
-      resetAfterDone: (delayAfterTriggered: number) => void,
-    }
-  ) => Promise<void>, overrides?: TriggerActionOverrides | undefined) {
-    // if there is one already running
-    await this.memeboxApi.actionActiveState.waitUntilDoneAsync(this.actionId, this.screenId);
-
-    const newOverrides = merge({}, overrides, {
-        screenMedia: {
-          visibility: VisibilityEnum.Toggle
-        }
-      });
-
-    console.info( { newOverrides });
-
-    this.memeboxApi.actionTriggerEventBus.triggerMedia({
-      id: this.actionId,
-      origin: TriggerClipOrigin.Scripts,
-      originId: this.memeboxApi.scriptId,
-      targetScreen: this.screenId,
-      overrides: newOverrides,
-      useOverridesAsBase: true
-    })
-
-    let resetAfterDone = 0;
-
-    await executionFunction({
-      reset: () => this.updateScreenOptions(null),
-      resetAfterDone: (delayAfterTriggered) => resetAfterDone  = delayAfterTriggered
-    });
-
-    await this.trigger(null);
-
-    if (resetAfterDone) {
-      await timeoutAsync(resetAfterDone);
-
-      this.updateScreenOptions(null)
-    }
-  }
-}
 
 // todo multiple actions
